requirements.txt
pandas
numpy
gunicorn
python-dotenv
flask
Flask-SQLAlchemy